+++ vehicle/autoware_accel_brake_map_calibrator/media/statistics_sample.png
