+++ vehicle/autoware_accel_brake_map_calibrator/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>accel_brake_map_calibrator</name>
+  <name>autoware_accel_brake_map_calibrator</name>
   <version>0.1.0</version>
   <description>The accel_brake_map_calibrator</description>
   <maintainer email="tomoya.kimura@tier4.jp">Tomoya Kimura</maintainer>
