+++ vehicle/autoware_accel_brake_map_calibrator/src/accel_brake_map_calibrator_node.cpp
@@ -26,7 +26,7 @@
 #include <utility>
 #include <vector>
 
-namespace accel_brake_map_calibrator
+namespace autoware::accel_brake_map_calibrator
 {
 
 AccelBrakeMapCalibrator::AccelBrakeMapCalibrator(const rclcpp::NodeOptions & node_options)
@@ -213,6 +213,7 @@ AccelBrakeMapCalibrator::AccelBrakeMapCalibrator(const rclcpp::NodeOptions & nod
     std::bind(&AccelBrakeMapCalibrator::callbackVelocity, this, _1));
   steer_sub_ = create_subscription<SteeringReport>(
     "~/input/steer", queue_size, std::bind(&AccelBrakeMapCalibrator::callbackSteer, this, _1));
+  std::cerr << "call bake set" << std::endl;
   if (accel_brake_value_source_ == ACCEL_BRAKE_SOURCE::STATUS) {
     actuation_status_sub_ = create_subscription<ActuationStatusStamped>(
       "~/input/actuation_status", queue_size,
@@ -286,15 +287,14 @@ void AccelBrakeMapCalibrator::timerCallback()
     "map updating... count: " << update_success_count_ << " / " << update_count_ << "\n\t"
                               << "lack_of_data_count: " << lack_of_data_count_ << "\n\t"
                               << " failed_to_get_pitch_count: " << failed_to_get_pitch_count_
+                              << "\n\t" << "too_large_pitch_count: " << too_large_pitch_count_
+                              << "\n\t" << " too_low_speed_count: " << too_low_speed_count_
+                              << "\n\t" << "too_large_steer_count: " << too_large_steer_count_
+                              << "\n\t" << "too_large_jerk_count: " << too_large_jerk_count_
+                              << "\n\t" << "invalid_acc_brake_count: " << invalid_acc_brake_count_
                               << "\n\t"
-                              << "too_large_pitch_count: " << too_large_pitch_count_ << "\n\t"
-                              << " too_low_speed_count: " << too_low_speed_count_ << "\n\t"
-                              << "too_large_steer_count: " << too_large_steer_count_ << "\n\t"
-                              << "too_large_jerk_count: " << too_large_jerk_count_ << "\n\t"
-                              << "invalid_acc_brake_count: " << invalid_acc_brake_count_ << "\n\t"
                               << "too_large_pedal_spd_count: " << too_large_pedal_spd_count_
-                              << "\n\t"
-                              << "update_fail_count_: " << update_fail_count_ << "\n");
+                              << "\n\t" << "update_fail_count_: " << update_fail_count_ << "\n");
 
   /* valid check */
 
@@ -304,7 +304,14 @@ void AccelBrakeMapCalibrator::timerCallback()
     !delayed_accel_pedal_ptr_ || !delayed_brake_pedal_ptr_) {
     // lack of data
     RCLCPP_WARN_STREAM_THROTTLE(
-      get_logger(), *get_clock(), 5000, "lack of topics (twist, steer, accel, brake)");
+      get_logger(), *get_clock(), 1000, "lack of topics (twist, steer, accel, brake)");
+    std::cerr << "!twist_ptr_" << !twist_ptr_ << std::endl;
+    std::cerr << "!steer_ptr_" << !steer_ptr_ << std::endl;
+    std::cerr << "!accel_pedal_ptr_" << !accel_pedal_ptr_ << std::endl;
+    std::cerr << "!brake_pedal_ptr_" << !brake_pedal_ptr_ << std::endl;
+    std::cerr << "!delayed_accel_pedal_ptr_" << !delayed_accel_pedal_ptr_ << std::endl;
+    std::cerr << "!delayed_brake_pedal_ptr_" << !delayed_brake_pedal_ptr_ << std::endl;
+
     lack_of_data_count_++;
     return;
   }
@@ -498,6 +505,7 @@ void AccelBrakeMapCalibrator::callbackVelocity(const VelocityReport::ConstShared
 
 void AccelBrakeMapCalibrator::callbackSteer(const SteeringReport::ConstSharedPtr msg)
 {
+  std::cerr << "AccelBrakeMapCalibrator::callbackSteer" << std::endl;
   debug_values_.data.at(CURRENT_STEER) = msg->steering_tire_angle;
   steer_ptr_ = msg;
 }
@@ -541,6 +549,7 @@ void AccelBrakeMapCalibrator::callbackActuation(
 void AccelBrakeMapCalibrator::callbackActuationCommand(
   const ActuationCommandStamped::ConstSharedPtr msg)
 {
+  std::cerr << "AccelBrakeMapCalibrator::callbackActuationCommand" << std::endl;
   const auto header = msg->header;
   const auto accel = msg->actuation.accel_cmd;
   const auto brake = msg->actuation.brake_cmd;
@@ -550,6 +559,7 @@ void AccelBrakeMapCalibrator::callbackActuationCommand(
 void AccelBrakeMapCalibrator::callbackActuationStatus(
   const ActuationStatusStamped::ConstSharedPtr msg)
 {
+  std::cerr << "AccelBrakeMapCalibrator::callbackActuationStatus" << std::endl;
   const auto header = msg->header;
   const auto accel = msg->status.accel_status;
   const auto brake = msg->status.brake_status;
@@ -1033,8 +1043,7 @@ bool AccelBrakeMapCalibrator::updateEachValOffset(
   RCLCPP_DEBUG_STREAM(
     get_logger(), "index: " << ped_idx << ", " << vel_idx
                             << ": map_offset_ = " << map_offset_vec_.at(ped_idx).at(vel_idx)
-                            << " -> " << map_offset << "\t"
-                            << " covariance = " << covariance);
+                            << " -> " << map_offset << "\t" << " covariance = " << covariance);
 
   /* input calculated result and update map */
   map_offset_vec_.at(ped_idx).at(vel_idx) = map_offset;
@@ -1062,8 +1071,7 @@ void AccelBrakeMapCalibrator::updateTotalMapOffset(const double measured_acc, co
   map_offset_ = map_offset_ + coef * error_map_offset;
 
   RCLCPP_DEBUG_STREAM(
-    get_logger(), "map_offset_ = " << map_offset_ << "\t"
-                                   << "covariance = " << covariance_);
+    get_logger(), "map_offset_ = " << map_offset_ << "\t" << "covariance = " << covariance_);
 
   /* update map */
   for (std::size_t ped_idx = 0; ped_idx < update_accel_map_value_.size() - 1; ped_idx++) {
@@ -1691,4 +1699,4 @@ void AccelBrakeMapCalibrator::addLogToCSV(
             << ", " << new_accel_mse << "," << rmse_rate << std::endl;
 }
 
-}  // namespace accel_brake_map_calibrator
+}  // namespace autoware::accel_brake_map_calibrator
