+++ vehicle/autoware_accel_brake_map_calibrator/test/plot_csv_client_node.py
@@ -82,7 +82,7 @@ def main(args=None):
 
 
 if __name__ == "__main__":
-    package_path = get_package_share_directory("accel_brake_map_calibrator")
+    package_path = get_package_share_directory("autoware_accel_brake_map_calibrator")
     parser = argparse.ArgumentParser()
     parser.add_argument(
         "-d", "--default-map-dir", default=None, type=str, help="directory of default map"
